Adds Deploy script to push prod build to GitHub pages docs folder
Adds script that creates a production build, moves it to the docs folder, and creates a commit of it. This allows us to create a build and update the GitHub pages website with the build.

Adds docs to .gitignore so that changes are not added to git by default except when deploying.